+++ src/youtube_automation/scripts/bulk_update_short_localizations.py
@@ -18,8 +18,8 @@ import logging
 import sys
 import time
 
+from youtube_automation.configuration import channel_dir, load_config
 from youtube_automation.utils.collection_paths import CollectionPaths
-from youtube_automation.utils.config import channel_dir, load_config
 from youtube_automation.utils.cost_tracker import log_quota
 from youtube_automation.utils.metadata_generator import build_short_localizations
 from youtube_automation.utils.youtube_service import get_youtube